% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kggplot.R
\name{kggplot}
\alias{kggplot}
\alias{kggplot.data.frame}
\alias{kggplot.kggplot}
\title{Plot data with ggplot}
\usage{
kggplot(x, ...)

\method{kggplot}{data.frame}(x, plot_type, vars, labels = NULL, ...)

\method{kggplot}{kggplot}(x, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Additional arguments to be passed to the plotting method.}

\item{vars}{\link{list} with the variables to be plotted.}

\item{labels}{\link{list} with the labels for the variables.}
}
\description{
This function is a generic method for plotting objects using ggplot2.
}
