% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\alias{kplot.default}
\alias{kplot.data.frame}
\title{Generic plot functions}
\usage{
kplot(x, ...)

\method{kplot}{default}(x, ...)

\method{kplot}{data.frame}(x, interactive = interactive(), output_format, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Additional arguments to be passed to the plotting method.}
}
\value{
A plot of the object \code{x}.
}
\description{
This function is a generic method for plotting objects.
}
\examples{
# Defaults
kplot(1:10)

}
