% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{createGgplotTheme}
\alias{createGgplotTheme}
\title{Create a Custom ggplot Theme}
\usage{
createGgplotTheme(template, ..., save = FALSE)
}
\arguments{
\item{template}{Character. The name of the theme template to apply (default:
"inrae").}

\item{...}{Additional arguments passed to methods.}

\item{save}{Logical. Whether to save the new theme (default: FALSE).}
}
\value{
A ggplot2 theme object.
}
\description{
Creates a new ggplot2 theme template. Not implemented yet.
}
