% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{getGgplotTheme}
\alias{getGgplotTheme}
\title{Get ggplot Theme}
\usage{
getGgplotTheme(template = "inrae", ...)
}
\arguments{
\item{template}{Character. The name of the theme template to apply (default:
"inrae").}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot2 theme object.
}
\description{
Retrieves a ggplot2 theme based on the specified template.
}
