% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{setPlotTheme}
\alias{setPlotTheme}
\alias{setPlotTheme.ggplot}
\title{Set Plot Theme for Various Plot Objects}
\usage{
setPlotTheme(x, ...)

\method{setPlotTheme}{ggplot}(x, template = "inrae", ...)
}
\arguments{
\item{x}{A plot object (e.g., a ggplot object).}

\item{...}{Additional arguments passed to methods.}

\item{template}{Character. The name of the theme template to apply (default:
"inrae").}
}
\value{
A plot object with the specified theme applied.
}
\description{
Generic function to set a plot theme for different types of plot objects.
}
