% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getValues.R
\name{getValues}
\alias{getValues}
\title{Get Aesthetic Value from Data}
\usage{
getValues(data, which)
}
\arguments{
\item{data}{A data frame or list containing the aesthetic values.}

\item{which}{A character string specifying the name of the aesthetic to retrieve.}
}
\value{
The value of the specified aesthetic from the data, or NULL if not found.
}
\description{
This function retrieves a specific aesthetic value from a data frame or list.
If the specified aesthetic is not found, it returns NULL.
}
